+++ Source/ImageCache.swift
@@ -6,8 +6,18 @@ public class ImageCache {
 
     // MARK: Class Types
 
-    /// A closure providing an optional image, boolean indicating true if the image was retrieved from the cache, and error.
-    public typealias ImageCompletion = (UIImage?, Bool, Error?) -> Void
+    /// Defines the source of the image.
+    public enum ImageSource {
+        /// The image was retrieved from the cache.
+        case cache
+        /// The image was retrieved from disk.
+        case localStorage
+        /// The image was retrieved from the remote source.
+        case remote
+    }
+
+    /// A closure providing an optional image, image source, and error.
+    public typealias ImageCompletion = (UIImage?, ImageSource, Error?) -> Void
 
     /// A closure to be returned by a remote source providing an optional image and error.
     public typealias RemoteImageCompletion = (UIImage?, Error?) -> Void
@@ -22,18 +32,26 @@ public class ImageCache {
     /// The object responding to delegate events.
     public weak var delegate: ImageCacheDelegate?
 
+    /// Provides an option to utilize local storage for files. If set to true, the images will be saved to the documents directory
+    /// in addition to being available in the cache. If false, files will only be added to the cache.
+    public var useLocalStorage: Bool = false
+
     // MARK: Internal Variables
 
     let cache: NSCache<AnyObject, UIImage>
+
+    let localFileController: LocalFileControllerProtocol
 
     // MARK: Init Methods
 
     init() {
         self.cache = NSCache<AnyObject, UIImage>()
+        self.localFileController = LocalFileController()
     }
 
-    init(cache: NSCache<AnyObject, UIImage>) {
+    init(cache: NSCache<AnyObject, UIImage>, localFileController: LocalFileControllerProtocol) {
         self.cache = cache
+        self.localFileController = localFileController
     }
 
     // MARK: Public Methods
@@ -41,15 +59,28 @@ public class ImageCache {
     /// Adds the provided image to the cache using the url as the key.
     ///
     /// - Parameters:
-    ///   - image: The image to cache.
-    ///   - url: The image's remote path.
-    public func cacheImage(_ image: UIImage?, forURL url: URL) {
+    ///   - image: The image to cache
+    ///   - url: The image's remote path
+    ///   - directory: The directory to save the image in if useLocalStorage is set to true.
+    public func cacheImage(_ image: UIImage?, forURL url: URL, directory: String? = nil) {
         guard let image = image else {
             return
         }
 
+        if useLocalStorage == true,
+            let imageName = imageNameFromURL(url) as? String {
+            localFileController.saveImage(image, fileName: imageName, directory: directory)
+        }
+
         let imageName = imageNameFromURL(url)
         cache.setObject(image, forKey: imageName)
+    }
+
+    /// Deletes the folder at the provided path.
+    ///
+    /// - Parameter directory: The directory to be deleted.
+    public func deleteDirectory(_ directory: String) {
+        localFileController.deleteDirectory(directory)
     }
 
     /// Removes all objects from the cache.
@@ -57,22 +88,30 @@ public class ImageCache {
         cache.removeAllObjects()
     }
 
-    /// Retrieves an images from the provided URL. This could be from the cache or from the remote source.
+    /// Retrieves an image from the provided URL. This could be from the cache or from the remote source.
     ///
     /// - Parameters:
     ///   - url: The URL for the image.
-    ///   - skipCache: If true, the cached is not checked first, the image is retrieved from the remote source and added to the cache.
+    ///   - directory: The directory of the image if useLocalStorage is set to true. This path will be appended to the documents directory path.
+    ///   - skipCache: If true, the cache is not checked first, the image is retrieved from the remote source and added to the cache.
     ///   - completion: The image, a boolean indicating true if the image was retrieved from the cache, and an error.
     /// - Returns: An optional data task if the image is retrieved from the remote source.
-    public func getImage(url: URL, skipCache: Bool = false, completion: @escaping ImageCompletion) -> URLSessionDataTask? {
+    public func getImage(url: URL, directory: String? = nil, skipCache: Bool = false, completion: @escaping ImageCompletion) -> URLSessionDataTask? {
         if let image = retrieveFromCache(url: url), skipCache == false {
-            completion(image, true, nil)
+            completion(image, .cache, nil)
+            return nil
+        }
+
+        if useLocalStorage == true,
+            let fileName = imageNameFromURL(url) as? String,
+            let image = localFileController.getImage(imageName: fileName, directory: directory) {
+            completion(image, .localStorage, nil)
             return nil
         }
 
         return delegate?.loadImageAtURL(url, completion: { [weak self] (image, error) in
-            self?.cacheImage(image, forURL: url)
-            completion(image, false, error)
+            self?.cacheImage(image, forURL: url, directory: directory)
+            completion(image, .remote, error)
         })
     }