+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 #### 1.x Releases
+- `1.1.0` Releases - [1.1.0](#110)
 - `1.0.x` Releases - [1.0.0](#100) | [1.0.1](#101)
+
+## [1.1.0](https://github.com/skladek/SKImageCache/releases/tag/1.1.0)
+
+#### Added
+- An option to allow saving files to disk in addition to the cache.
+- An option to delete local file directories from the image cache.
 
 ## [1.0.1](https://github.com/skladek/SKImageCache/releases/tag/1.0.1)